refactor: use parse argv
