+++ packages/core-common/src/node/cli.ts
@@ -1,20 +1,7 @@
-/**
- * 一个方便使用的 argv 提取工具
- * > mri("node index.js -a -b=1 -c 2".split())
- *
- * {
- *   _: [
- *     'node',
- *     'index.js'
- *   ],
- *   a: true,
- *   b: 1,
- *   c: 2
- * }
- */
+/* eslint-disable no-console */
 
-import mri from 'mri';
+import { parseArgv } from '@opensumi/ide-utils/lib/argv';
 
-const argv = mri(process.argv);
+const argv = parseArgv(process.argv);
 
 export { argv };
