+++ packages/remote-cli/package.json
@@ -20,7 +20,8 @@
   "dependencies": {
     "chalk": "^4.1.2",
     "got": "^12.1.0",
-    "mri": "^1.2.0"
+    "mri": "^1.2.0",
+    "@opensumi/ide-utils": "2.20.2"
   },
   "devDependencies": {
     "@opensumi/ide-dev-tool": "^1.3.1"
