+++ tools/electron/src/node/server.ts
@@ -1,14 +1,12 @@
 /* eslint-disable no-console */
 import net from 'net';
 
-import mri from 'mri';
-
 import { Deferred } from '@opensumi/ide-core-common';
 import { DEFAULT_OPENVSX_REGISTRY } from '@opensumi/ide-core-common/lib/const';
 import { IServerAppOpts, ServerApp, NodeModule } from '@opensumi/ide-core-node';
+import { parseArgv } from '@opensumi/ide-utils/lib/argv';
 
-const _argv = process.argv.slice(2);
-const argv = mri(_argv);
+const argv = parseArgv(process.argv.slice(2));
 
 export async function startServer(arg1: NodeModule[] | Partial<IServerAppOpts>) {
   const deferred = new Deferred<net.Server>();
